Use more appropriate conventions for python module
Name in setup.py is "pulpstatus", top-level python module has the same
name, and the WSGI callable is named "application" (as preferred by
gunicorn conventions).

Useless wsgi.py has also been removed, and the Makefile was updated to
use gunicorn so that the same server is used locally as in OpenShift.